+++ app/assets/scripts/app.js
@@ -43,6 +43,8 @@
             c6UrlMakerProvider.location(c6Defines.kCollateralUrl,'collateral');
             c6UrlMakerProvider.location(c6Defines.kApiUrl,'api');
             c6UrlMakerProvider.location(c6Defines.kProtocol + '/', 'protocol');
+            c6UrlMakerProvider.location(c6Defines.kEnvUrlRoot,'envroot');
+            window.console.log('KENV:',c6Defines.kEnvUrlRoot);
         }])
         .config(['VASTServiceProvider', 'VPAIDServiceProvider',
         function( VASTServiceProvider, VPAIDServiceProvider ) {
@@ -71,6 +73,13 @@
                 return Math.round(isNaN(input) ? 0 : (input * 100)) + '%';
             };
         })
+        .filter('envroot', ['c6UrlMaker',
+        function              ( c6UrlMaker ) {
+            return function(url) {
+                window.console.log('envurl:',c6UrlMaker(url,'envroot'));
+                return url && c6UrlMaker(url.replace(/^\//,''), 'envroot');
+            };
+        }])
         .filter('collateral', ['c6UrlMaker',
         function              ( c6UrlMaker ) {
             return function(url) {
@@ -272,7 +281,7 @@
 
             function setMode(obj, data) {
                 var device = data.profile.device,
-                    mode = data.experience.mode;
+                    mode = data.experience.data.mode;
 
                 obj.mode = (device === 'phone') ? 'mobile' : (mode || 'full');
             }